Create 'create new color palette' feature
Created color-palette component that contains four input field
components:
- color-input-one
- color-input-two
- color-input-three
- color-input-four

The color-palette component also contains a submit button, meant for
a user to click on once all input fields are filled in.

Defined 'submitPalette' action in color-palette.js. The action currently
contains a console.log in order to confirm that the 'palette' object
returned in the console does indeed contain the user inputs for each of
the fields.

Next step: Write code to persist the user-generated palette in the
database.