Rebuild 'create new palette' feature
Previous 'create new palette' responded with '422 unprocessable entity'
error in console. This is because, even though the 'post' action is
sent, the user-inputted data could not be persisted to the database
because my-application.js could not properly access the user-inputted
data.

Generated a 'create' router, which created:
- routes/create.js
- templates/create.hbs
- test file

Generated a 'palette-create' component, which is placed inside the
'create' template.

Send the 'createPalette' action up to the 'create' router so that
the user-inputted data will get persisted to the database.

Now, the create feature will successfully create a new palette.