Display all palette colors in 'My Palletes' view
Create a computed property for each color of a palette in the the js
file of the component, palettes-all:
- colorBg1
- colorBg2
- colorBg3
- colorBg4

Using a computed property for each color is a workaround for applying
style directly on each color in hbs file. Otherwise, I get the following
warning: "WARNING: Binding style attributes may introduce cross-site
scripting vulnerabilities; please ensure that values being bound are
properly escaped."

With the workaround, the computed property for each color will return
the 'html safe' background-color styling.
More on 'html safe' from emberjs.com: "Mark a string as safe for
unescaped output with Ember templates. If you return HTML from a helper,
use this function to ensure Ember's rendering layer does not escape the
HTML."

Then in palettes-all.hbs, set background-color (inline-style) of each
color as the computed property for the corresponding color.

Now, when users visit the 'My Palettes' view, all colors of the all
palettes get displayed as background-color for each hex code, which is
still visible. Also, when users edit a color, click submit, and is taken
back to the 'My Palettes' page, the colors changed will be updated.
Clicking 'back' without submission will still display the original
colors as expected as well.