Redirect /create to sign in if user not signed in
Until now, user was able to view /palettes/create page if added
'palettes/create' to homepage URL (Although a user still cannot create a
new palette when not signed in).

In order to provide an error message and redirect to the sign-in page
like it does for the /palettes route, I hit the backend API in the
router model so that it triggers the error action in application.js.

Note: /palettes route already hits the backend API in its model, so
that is why it already provides an error message and redirects.