Update edited palette without binding
In components/palette-edit-form.js, set user input to 'editedPalette'
instead of 'palette' so that user input will not initially be bound to
original palette.

In templates/components/palette-edit-form.hbs, edit value=palette.colorx
to value=editedPalette.colorx to bind user-inputted data to
editedPalette instead of original 'palette'.

If instead set value=palette.color1 in palette-edit-form.hbs, then
this.get('palette') (original palette saved in database) will
automatically display user-inputted data colors,
although not saved in DB. In this case, this is not desired due to the
fact that users could click on 'My Palettes' button before submitting
edited palette. The colors in palettes on My Palettes page will update
EVEN IF user had not submitted the edited palette).

Now, when user clicks on 'submit edited palette', the 'palette' with
new colors will be sent to palette-edit router to be saved to the
database.